Fix FileWatcher skipping DELETE events
The FileWatcher was skipping DELETE events, except on very weird cases
where the file was deleted and re-created just before the event was
triggered by awatch().

This also extends the integration test to test both situations (delete
events when a file exist and when it doesn't).

Because there is usually a considerable delay between a change in the
filesystem happens and awatch() emits an event it is difficult to decide
if we should filter events based on the file existing or not. For
example, if a create event is received but before is triggered the file
was removed, should we emit the create event or not? Or the other way
around, should we emit a delete event if a file was created again before
the event was triggered.

It is probably best to leave this decision to the user, and let them
deal with races.

+++ tests/utils/test_integration.py
@@ -46,8 +46,11 @@ async def test_file_watcher(tmp_path: pathlib.Path) -> None:
 
 
 @pytest.mark.integration
-async def test_file_watcher_change_types(tmp_path: pathlib.Path) -> None:
+async def test_file_watcher_deletes(tmp_path: pathlib.Path) -> None:
     """Ensure file watcher is returning paths only on the DELETE change.
+
+    Also ensures that DELETE events are sent even if the file was recreated and even if
+    the file doesn't exist.
 
     Args:
         tmp_path (pathlib.Path): A tmp directory to run the file watcher on.
@@ -63,18 +66,44 @@ async def test_file_watcher_change_types(tmp_path: pathlib.Path) -> None:
         deletion_timer=Timer.timeout(timedelta(seconds=0.25)),
         watcher=file_watcher,
     )
-    number_of_deletes = 0
     number_of_write = 0
+    number_of_deletes = 0
+    number_of_events = 0
+    # We want to write to a file and then removed, and then write again (create it
+    # again) and remove it again and then stop.
+    # Because awatch takes some time to get notified by the OS, we need to stop writing
+    # while a delete was done, to make sure the file is not recreated before the
+    # deletion event arrives.
+    # For the second round of writes and then delete, we allow writing after the delete
+    # was done as an extra test.
+    #
+    # This is an example timeline for this test:
+    #
+    # |-----|--.--|-----|---o-|-----|--.--|-----|--o--|-----|-----|-----|-----|-----|
+    # W     W  D            E W     W  D  W     W  E
+    #
+    # Where:
+    # W: Write
+    # D: Delete
+    # E: FileWatcher Event
     while await select.ready():
         if msg := select.write_timer:
+            if number_of_write >= 2 and number_of_events == 0:
+                continue
             filename.write_text(f"{msg.inner}")
             number_of_write += 1
         elif _ := select.deletion_timer:
+            # Avoid removing the file twice
+            if not pathlib.Path(filename).is_file():
+                continue
             os.remove(filename)
-        elif _ := select.watcher:
             number_of_deletes += 1
-            break
+        elif _ := select.watcher:
+            number_of_events += 1
+            if number_of_events >= 2:
+                break
 
-    assert number_of_deletes == 1
+    assert number_of_deletes == 2
     # Can be more because the watcher could take some time to trigger
-    assert number_of_write >= 2
+    assert number_of_write >= 3
+    assert number_of_events == 2
